requirements.txt
requests==2.32.4
# Pin urllib3 to <2 to avoid NotOpenSSLWarning on macOS system Python with LibreSSL
urllib3<2.0
requests-cache==1.2.1
beautifulsoup4==4.12.3
lxml==5.1.0
qdrant-client==1.11.1
sentence-transformers==3.0.1
PyYAML==6.0.2
python-json-logger==2.0.7
python-dotenv==1.0.1
APScheduler==3.10.4
# torch left unpinned to allow compatible version for current Python (e.g., 2.7.x for Py3.13)
